requirements.txt
fastapi>=0.95
uvicorn
pydantic
python-dotenv
pytest
openai
